weer terugzetten double-int

+++ src/Config.h
@@ -31,7 +31,7 @@ public:
 	int SimpleLengthFactor;
 	int numStab;
 	int Hifactor;
-	double Minfactor;
+	int Minfactor;
 	double Hcrit_global;
 	int transport_eq;
 	int alpha_varies;
